Change style in MainActivity, and broke up methods into better chunks
