+++ app/src/main/java/mccode/qdup/QueryModels/External_ids.java
@@ -1,43 +1,43 @@
 
-package mccode.qdup.models;
+package mccode.qdup.QueryModels;
 
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
 
-public class External_urls implements Serializable
+public class External_ids implements Serializable
 {
 
-    private String spotify;
+    private String isrc;
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-    private final static long serialVersionUID = -6992500396595992958L;
+    private final static long serialVersionUID = 6449473715479390211L;
 
     /**
      * No args constructor for use in serialization
      * 
      */
-    public External_urls() {
+    public External_ids() {
     }
 
     /**
      * 
-     * @param spotify
+     * @param isrc
      */
-    public External_urls(String spotify) {
+    public External_ids(String isrc) {
         super();
-        this.spotify = spotify;
+        this.isrc = isrc;
     }
 
-    public String getSpotify() {
-        return spotify;
+    public String getIsrc() {
+        return isrc;
     }
 
-    public void setSpotify(String spotify) {
-        this.spotify = spotify;
+    public void setIsrc(String isrc) {
+        this.isrc = isrc;
     }
 
-    public External_urls withSpotify(String spotify) {
-        this.spotify = spotify;
+    public External_ids withIsrc(String isrc) {
+        this.isrc = isrc;
         return this;
     }
 
@@ -49,7 +49,7 @@ public class External_urls implements Serializable
         this.additionalProperties.put(name, value);
     }
 
-    public External_urls withAdditionalProperty(String name, Object value) {
+    public External_ids withAdditionalProperty(String name, Object value) {
         this.additionalProperties.put(name, value);
         return this;
     }
