+++ app/src/main/java/mccode/qdup/QueryModels/Message.java
@@ -1,4 +1,4 @@
-package mccode.qdup.models;
+package mccode.qdup.QueryModels;
 
 import java.io.Serializable;
 import java.util.HashMap;
