+++ app/src/main/java/mccode/qdup/ServerActivity.java
@@ -42,9 +42,9 @@ import mccode.qdup.Utils.Listeners.SearchListener;
 import mccode.qdup.Utils.QueueView.HostRecyclerListAdapter;
 import mccode.qdup.Utils.Server.ServerListener;
 import mccode.qdup.Utils.Server.ServerWriter;
-import mccode.qdup.models.Item;
-import mccode.qdup.models.ResponseWrapper;
-import mccode.qdup.models.TrackResponse;
+import mccode.qdup.QueryModels.Item;
+import mccode.qdup.QueryModels.ResponseWrapper;
+import mccode.qdup.QueryModels.TrackResponse;
 
 import static mccode.qdup.MainActivity.serverCode;
 import static mccode.qdup.MainActivity.musicPlayer;
