+++ app/src/main/java/mccode/qdup/Utils/QueueView/RequesterRecyclerListAdapter.java
@@ -15,10 +15,7 @@ import java.util.List;
 
 import mccode.qdup.R;
 import mccode.qdup.RequesterActivity;
-import mccode.qdup.ServerActivity;
-import mccode.qdup.Utils.Messaging.Message;
-import mccode.qdup.Utils.Messaging.MessageCode;
-import mccode.qdup.models.Item;
+import mccode.qdup.QueryModels.Item;
 
 /**
  * Author: Connor McAuliffe
